Allow for caching of submodules.
Added `git_submodule_cache_all` to initialze a cache of submodules on the
repository so that operations looking up N submodules are O(N) and not O(N^2).
Added a `git_submodule_clear_cache` function to remove the cache.

Also optimized the function that loads all submodules as it was itself O(N^2)
w.r.t the number of submodules, having to loop through the `.gitmodules` file
once per submodule.  I changed it to process the `.gitmodules` file once, into
a map.